+++ lib/src/header_parser/sub_parsers/functiondecl_parser.dart
@@ -36,8 +36,9 @@ List<Func>? parseFunctionDeclaration(clang_types.CXCursor cursor) {
 
     final rt = _getFunctionReturnType(cursor);
     final parameters = _getParameters(cursor, funcName);
-
-    if (clang.clang_Cursor_isFunctionInlined(cursor) != 0) {
+    if (clang.clang_Cursor_isFunctionInlined(cursor) != 0 &&
+        clang.clang_Cursor_getStorageClass(cursor) !=
+            clang_types.CX_StorageClass.CX_SC_Extern) {
       _logger.fine('---- Removed Function, reason: inline function: '
           '${cursor.completeStringRepr()}');
       _logger.warning(